switched over to $ILRepackCLI = $rootDir + "tools\ILRepack.MSBuild.Task.2.0.0\tools\ILRepack.exe"
1. I don't need to call PDB2PDB anymore
2. It has solve the Visual Studio issue where it was stepping the wrong code.

+++ tools/ensurePortablePDB.ps1
@@ -52,27 +52,30 @@ foreach ($item in $jsonObj.assembliesToMerge) {
     }
 }
 
-foreach ($item in $jsonObj.assembliesToMerge) {
-    $assembly = $item.assembly;
-    $assemblyPath = $targetDir + $assembly + ".dll"
-    $version = Get-ChildItem $assemblyPath | Select-Object -ExpandProperty VersionInfo
+if ($false) {
+    
+    foreach ($item in $jsonObj.assembliesToMerge) {
+        $assembly = $item.assembly;
+        $assemblyPath = $targetDir + $assembly + ".dll"
+        $version = Get-ChildItem $assemblyPath | Select-Object -ExpandProperty VersionInfo
   
-    $pdbPath = $targetDir + $assembly + ".pdb"
+        $pdbPath = $targetDir + $assembly + ".pdb"
   
-    if ((Test-Path $pdbPath) ) {
+        if ((Test-Path $pdbPath) ) {
 
-        & $pdb2PDBCLI $assemblyPath
-        if (-not $?) {
-            write-output "Something went wrong with pdb2pdb"
-            $ExitCode = "-1"
-            Exit $ExitCode
+            & $pdb2PDBCLI $assemblyPath
+            if (-not $?) {
+                write-output "Something went wrong with pdb2pdb"
+                $ExitCode = "-1"
+                Exit $ExitCode
+            }
+            $portablePDBPath = $pdbPath + "2" #appends a 2
+            Remove-Item $pdbPath
+            Rename-Item -Path $portablePDBPath -NewName $pdbPath
+
         }
-        $portablePDBPath = $pdbPath + "2" #appends a 2
-        Remove-Item $pdbPath
-        Rename-Item -Path $portablePDBPath -NewName $pdbPath
-
-    }
    
+    }
 }
